Disable building of the compiler by default.
**Why Is This Necessary**

1. The compiler is currently being developed and is not useful yet.
2. Building the compiler requires the builder to have llvm-6.0+ installed correctly + pulls
   in other, otherwise unnecessary dependencies.

This offers low return on investment of time and harms ergonomics and so the compiler
should not be built by a naive `stack build` on master. This should ultimately change in
the future.

**If the folders and files are still there how does this work?**

Commenting out the directory in the stack.yml means that stack doesn't know about it
and so won't go looking for the package.yml in the compiler/ directory.