Prevent overlapping instances
Support for overlapping instances was a bit sketchy, and the
most-recently defined instance was used, rather than the most
specific, and the order in which instances was defined determined
whether or not a duplicate instance error would occur.

Support for overlapping instances is desirable to have in the future,
but we must have a robust system for deciding how to favour one
instance over another (e.g., most 'specific'), and when the overlap is
unresolvable.