Bump text-replace version to fix "stack build"
Stack is configured to use lts-15.13 and `text-replace-0.0.0.4`. This
leads to an error when running `stack build`:

```
Error: While constructing the build plan, the following exceptions were encountered:

In the dependencies for text-replace-0.0.0.4:
    base-4.13.0.0 from stack configuration does not match >=4.9 && <4.13  (latest matching version is 4.12.0.0)
    optparse-applicative-0.15.1.0 from stack configuration does not match >=0.12.1.0 && <0.15  (latest matching version
                                  is 0.14.3.0)
needed due to granule-frontend-0.8.1.0 -> text-replace-0.0.0.4

Some different approaches to resolving this:

  * Set 'allow-newer: true' in /home/john/.stack/config.yaml to ignore all version constraints and build anyway.

  * Build requires unattainable version of base. Since base is a part of GHC, you most likely need to use a different
    GHC version with the matching base.

Plan construction failed.
```

`text-replace-0.0.0.6` bumps the dependencies for `base` and
`optparse-applicative` to versions that are available in lts-15.13